Reworked Player again
