+++ Assets/Scripts/Interfaces/IHackable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4adb11e1f8406da4aa81bfe5ffcf726b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
