+++ Assets/Scripts/Player/PlayerVariables.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 89545a51fd86038409392f0a99f9856e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
